Rewrite check for square brackets in exec form
This was initially committed on Jan 8.
The current implementation simplifies it by using JSON.parse.
It also simplifies the alerts by just saying that the array is invalid, instead of specifying the extra brackets.
It also is more comprehensive in terms of cases covered.
The added unit tests exemplify this.